Create separate handler for clusters
